requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
tqdm>=4.66.0
rich>=13.7.0 